requirements.txt
opencv-python
# torch torchvision --extra-index-url https://download.pytorch.org/whl/cu118 #install this manually
ultralytics
numpy
onnxruntime-gpu
insightface
qdrant-client